Migreate iframed inline-styles test to playwright
